Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.4.1
firebase-admin==6.4.0
python-dotenv==1.0.1
pydantic==2.6.1
google-auth==2.28.0
requests==2.31.0
